+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Change Log
 
-## X.X.X - XXXX-XX-XX
+## 0.2.0- 2017-09-27
 
-* TODO
-    * Add antrun plugin to clear pojo builder.
+* Enabled Spring Boot.
+* Added `.gitignore` in archetype. 
  
 ## 0.1.0 - 2015-11-04